+++ NEWS.md
@@ -2,12 +2,13 @@
 
 ## New Features
 
-- Argument `ell` now overloaded:
+- Allow argument `ell` to be shorter than covariate dimension. In this case,
+  `ell` specifies which subset of covariates to compute standard errors for.
 
 ## Minor improvements and fixes
 
 - Use `collapse::fsum` instead of `tapply` calls to improve speed
-
+- Check that covariates are not collinear
 
 # dfadjust 1.0.5